forward username & bad link handling

+++ src/bot.py
@@ -18,29 +18,37 @@ def download_and_send(message):
     text = message.text
     message_id = message.message_id
     thread_id = message.message_thread_id
-    username = message.from_user.username
+    if (sender:=message.forward_from):
+        username = sender.username
+    else:
+        username = message.from_user.username 
     bot.delete_message(chat_id, message_id)
     bot_message = bot.send_message(chat_id=chat_id,
                                    message_thread_id=thread_id,
                                    text='ща будет рилс...')
-    pattern = r'https://www.instagram.com/reel/(.*)/.*'
-    shortcode = re.match(pattern, text).group(1)
-    try:
-        post = instaloader.Post.from_shortcode(L.context, shortcode)
-        L.download_post(post, target=target_directory)
-        for file in os.listdir(target_directory):
-            if file.endswith('.mp4'):
-                # bot.edit_message_media(chat_id=chat_id, message_id=bot_message.message_id, media=open(f'reels/{file}', 'rb'))
-                bot.send_video(chat_id=chat_id,
-                               message_thread_id=thread_id,
-                               video=open(f'{target_directory}/{file}', 'rb'),
-                               caption=f'рилс от @{username}')
-                # bot.edit_message_text(chat_id=chat_id, message_id=bot_message.message_id, text='')
-                bot.delete_message(chat_id, bot_message.message_id)
-            os.remove(f'reels/{file}')
-    except instaloader.exceptions.InstaloaderException as e:
+    matched = re.match(r'https://www.instagram.com/reel/(.*)/.*', text)
+    if not matched:
         bot.edit_message_text(chat_id=chat_id,
                               message_id=bot_message.message_id,
-                              text=f'рилса не будет :(\nошибка: {e}')
+                              text=f'ты кого наебать пытаешься?')
+    else:
+        shortcode = matched.group(1)
+        try:
+            post = instaloader.Post.from_shortcode(L.context, shortcode)
+            L.download_post(post, target=target_directory)
+            for file in os.listdir(target_directory):
+                if file.endswith('.mp4'):
+                    # bot.edit_message_media(chat_id=chat_id, message_id=bot_message.message_id, media=open(f'reels/{file}', 'rb'))
+                    bot.send_video(chat_id=chat_id,
+                                message_thread_id=thread_id,
+                                video=open(f'{target_directory}/{file}', 'rb'),
+                                caption=f'рилс от @{username}')
+                    # bot.edit_message_text(chat_id=chat_id, message_id=bot_message.message_id, text='')
+                    bot.delete_message(chat_id, bot_message.message_id)
+                os.remove(f'reels/{file}')
+        except instaloader.exceptions.InstaloaderException as e:
+            bot.edit_message_text(chat_id=chat_id,
+                                message_id=bot_message.message_id,
+                                text=f'рилса не будет :(\nошибка: {e}')
 
 bot.polling()
